Distance walked in my profile should now update when the user moves

+++ app/src/main/java/ntnu/imt3673/android_geocache/AchievementHandler.java
@@ -53,7 +53,7 @@ public class AchievementHandler {
         return achievementsQualified;
     }
 
-    public static ArrayList<Integer> meetDistanceCriteria(int distanceNum){
+    public static ArrayList<Integer> meetDistanceCriteria(double distanceNum){
         ArrayList<Achievement> ret = returnAchievementList();
         ArrayList<Integer> achievementsQualified = new ArrayList<>();
